requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/prod.in
#
appdirs==1.4.4
    # via fs
authlib==1.1.0
    # via simple-salesforce
certifi==2022.9.24
    # via
    #   requests
    #   snowfakery
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via
    #   requests
    #   snowfakery
click==8.1.3
    # via
    #   -r requirements/prod.in
    #   snowfakery
commonmark==0.9.1
    # via rich
cryptography==38.0.1
    # via
    #   -r requirements/prod.in
    #   authlib
    #   pyjwt
defusedxml==0.7.1
    # via -r requirements/prod.in
docutils==0.16
    # via -r requirements/prod.in
faker==15.0.0
    # via
    #   -r requirements/prod.in
    #   snowfakery
fs==2.4.16
    # via -r requirements/prod.in
github3-py==3.2.0
    # via -r requirements/prod.in
greenlet==1.1.3.post0
    # via snowfakery
gvgen==1.0
    # via snowfakery
idna==3.4
    # via
    #   requests
    #   snowfakery
importlib-metadata==5.0.0
    # via keyring
jinja2==3.1.2
    # via
    #   -r requirements/prod.in
    #   snowfakery
keyring==23.0.1
    # via -r requirements/prod.in
lxml==4.9.1
    # via -r requirements/prod.in
markupsafe==2.1.1
    # via
    #   -r requirements/prod.in
    #   jinja2
    #   snowfakery
psutil==5.9.3
    # via -r requirements/prod.in
pycparser==2.21
    # via cffi
pydantic==1.10.2
    # via
    #   -r requirements/prod.in
    #   snowfakery
pygments==2.13.0
    # via rich
pyjwt[crypto]==2.6.0
    # via
    #   -r requirements/prod.in
    #   github3-py
python-baseconv==1.2.2
    # via snowfakery
python-dateutil==2.8.2
    # via
    #   -r requirements/prod.in
    #   faker
    #   github3-py
    #   snowfakery
pytz==2022.5
    # via -r requirements/prod.in
pyyaml==6.0
    # via
    #   -r requirements/prod.in
    #   snowfakery
requests==2.28.1
    # via
    #   -r requirements/prod.in
    #   github3-py
    #   requests-futures
    #   robotframework-requests
    #   salesforce-bulk
    #   simple-salesforce
    #   snowfakery
requests-futures==1.0.0
    # via -r requirements/prod.in
rich==12.6.0
    # via -r requirements/prod.in
robotframework==6.0
    # via
    #   -r requirements/prod.in
    #   robotframework-lint
    #   robotframework-pabot
    #   robotframework-requests
    #   robotframework-seleniumlibrary
    #   robotframework-stacktrace
robotframework-lint==1.1
    # via -r requirements/prod.in
robotframework-pabot==2.7.0
    # via -r requirements/prod.in
robotframework-pythonlibcore==3.0.0
    # via robotframework-seleniumlibrary
robotframework-requests==0.9.3
    # via -r requirements/prod.in
robotframework-seleniumlibrary==5.1.3
    # via -r requirements/prod.in
robotframework-stacktrace==0.4.1
    # via robotframework-pabot
rst2ansi==0.1.5
    # via -r requirements/prod.in
salesforce-bulk==2.2.0
    # via -r requirements/prod.in
sarge==0.1.7.post1
    # via -r requirements/prod.in
selenium==3.141.0
    # via
    #   -r requirements/prod.in
    #   robotframework-seleniumlibrary
simple-salesforce==1.11.4
    # via
    #   -r requirements/prod.in
    #   salesforce-bulk
six==1.16.0
    # via
    #   fs
    #   python-dateutil
    #   salesforce-bulk
    #   snowfakery
snowfakery==3.4.0
    # via -r requirements/prod.in
sqlalchemy==1.4.41
    # via
    #   -r requirements/prod.in
    #   snowfakery
typing-extensions==4.4.0
    # via
    #   pydantic
    #   rich
    #   snowfakery
unicodecsv==0.14.1
    # via salesforce-bulk
uritemplate==4.1.1
    # via github3-py
urllib3==1.26.12
    # via
    #   requests
    #   selenium
    #   snowfakery
xmltodict==0.13.0
    # via -r requirements/prod.in
zipp==3.10.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
